Correcty types in minimal fee calculation

+++ src/CryptoNoteCore/Currency.cpp
@@ -417,7 +417,7 @@ namespace CryptoNote {
 		const uint64_t blocksInTwoYears = CryptoNote::parameters::EXPECTED_NUMBER_OF_BLOCKS_PER_DAY * 365 * 2;
 		const double gauge = double(0.25);
 		uint64_t minimumFee(0);
-		double dailyDifficultyMoore = dailyDifficulty / pow(2, height / blocksInTwoYears);
+		double dailyDifficultyMoore = dailyDifficulty / pow(2, static_cast<double>(height) / static_cast<double>(blocksInTwoYears));
 		double minFee = gauge * CryptoNote::parameters::COIN * static_cast<double>(avgHistoricalDifficulty) 
 			/ dailyDifficultyMoore * static_cast<double>(reward)
 			/ static_cast<double>(medianHistoricalReward);
